Refactored chart into own module
