+++ e_metrobus/navigation/views.py
@@ -1,12 +1,11 @@
-import plotly
-import plotly.graph_objects as go
 
 from django.views.generic import TemplateView
 from django.shortcuts import redirect
 
+from e_metrobus.navigation import chart
 from e_metrobus.navigation import constants
 
-from e_metrobus.navigation import widgets, utils
+from e_metrobus.navigation import widgets
 
 
 class NavigationView(TemplateView):
@@ -78,33 +77,5 @@ class PlotlyView(TemplateView):
 
     def get_context_data(self, **kwargs):
         context = super(PlotlyView, self).get_context_data(**kwargs)
-
-        colors = ["Gainsboro",] * 5
-        colors[2] = "black"
-
-        # Plotly Figure:
-        mobiles = ["Zu Fuß", "Fahrrad", "E-Bus", "Dieselbus", "PKW"]
-        fig = go.Figure([go.Bar(x=mobiles, y=[5, 5, 20, 30, 40], marker_color=colors)])
-        fig.layout.margin.t = 0
-        fig.layout.margin.b = 0
-        fig.layout.margin.l = 0
-        fig.layout.margin.r = 0
-        fig.layout.autosize = True
-        fig.layout.plot_bgcolor = "#fff"
-        fig.layout.yaxis.visible = False
-
-        fig.add_layout_image(
-            go.layout.Image(source="/static/images/icons/i_walk.svg", x=0, y=0.1)
-        )
-        fig.add_layout_image(
-            go.layout.Image(source="/static/images/icons/i_bus.svg", x=1, y=0.1)
-        )
-        fig.update_layout_images(
-            dict(
-                xref="x", yref="paper", sizex=0.2, sizey=0.2, xanchor="right", yanchor="bottom"
-            )
-        )
-
-        dj_fig = utils.DjangoFigure(fig, displayModeBar=False)
-        context["plotly"] = dj_fig
+        context["plotly"] = chart.get_mobility_figure([5, 5, 10, 30, 40])
         return context
